requirements.txt
slack-bolt
python-dotenv 